requirements.txt
pandas>=1.5.0
numpy>=1.21.0
lightgbm>=3.3.0
catboost>=1.2.0
scikit-learn>=1.0.0
PyYAML>=6.0
